Change css to highlight/gray when hover over any calendar date, even those with reminders.
It was hard to tell when you were hovering over a reminder when there were many events listed on the calendar widget, as days were not highlighting on hover if they had an event on them.

Also disabled clicking a day on the calendar header to change the start date.  Several users had done that inadvertenly and couldn't figure out what had happened.  This is not a common operation so just having to set it via the options is reasonable.