Create README.md
ReminderFox is an extension for Firefox and Thunderbird that displays and manages lists of date-based reminders and ToDo's.  ReminderFox does not seek to be a full-fledged calendar system. In fact, the target audience is anybody that simply wants to remember important dates (birthdays, anniversaries, etc) without having to run an entire calendar application.