Revert bitbucket change of removing reminderfox_smartFoxyBadge_bow/styles
The status bow toolbar icon was not displaying the number overlay.  This is because the
reminderfox_smartFoxyBadge_bow is a separate dom element, and the css styles are tied to that element.

To properly remove this, we need to attach a class to the smartFoxyBadge elements and tie the styles to the classname and
not the toolbar button element name itself.