Fix issue where clicking Firefox History toolbar button was showing icons only and not text when ReminderFox installed.
This was due to the fact that in toolbar.css we were disabling the text of the GLOBAL toolbar icons; looks like that must have been copied from some Firefox css long ago, as it has been that way for at least the last 3 years according to bitbucket source code history.

Removed the offending css and it fixed the issue, and seems to have no ill effect on any ReminderFox toolbar buttons.